pyproject.toml
[tool.poetry]
name = "ai-agents"
version = "0.1.0"
description = ""
authors = ["MAQuesada <malejandroquesada@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = "^21.10"
langchain = "^0.3.17"
langchain-openai = "^0.3.3"
langchain-google-genai = "^2.1.0"
python-dotenv = "^1.0.1"
langgraph = "^0.2.69"
composio-langchain = "^0.7.0"
psycopg = {extras = ["binary"], version = "^3.2.4"}
langgraph-checkpoint-postgres = "^2.0.13"
psycopg-pool = "^3.2.4"
pytz = "^2025.1"
composio = "^0.8.8"


[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
ruff = "^0.9.4"
pre-commit = "^4.1.0"
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
